+++ packages/mcap/src/McapReader.test.ts
@@ -2,8 +2,9 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import McapReader, { parseRecord } from "./McapReader";
+import McapReader from "./McapReader";
 import { MCAP_MAGIC, RecordType } from "./constants";
+import { parseRecord } from "./parse";
 
 function uint32LE(n: number): Uint8Array {
   const result = new Uint8Array(4);
